Fix ambuguos function call on unix builds

+++ demo/main.cpp
@@ -21,7 +21,7 @@ int main(int argc, char* argv[]){
     myParser.registerPositional<int>(
         "positional",
         "integer",
-        [&myParser](){ std::cout << "Entered value " << std::to_string(myParser.getValue<int>("positional")) << std::endl; },
+        [&myParser](){ std::cout << "Entered value " << std::to_string((int)myParser.getValue<int>("positional")) << std::endl; },
         [](const std::string& str)->int{ return std::stoi(str); },
         [value](const int& i)->bool{ bool valid = i < value ? true : false; return valid; },
         [](const std::exception& e){ std::cout << e.what(); }
